pyproject.toml
[tool.poetry]
name = "miur-backend"
version = "0.1.0"
description = ""
authors = ["Mateus Santos <santos.mateus@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask-SQLAlchemy = "^2.5.1"
connexion = {extras = ["swagger-ui"], version = "^2.14.0"}
python-dotenv = "^0.20.0"
Flask-Cors = "^3.0.10"
Flask-Migrate = "^3.1.0"
psycopg2 = "^2.9.3"
pycardano = "0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
retry = "^0.9.2"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
